requirements.txt
numpy
opencv-python
matplotlib
# hashlib
pycryptodome